requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0
